Se genero Menu de opciones a conversar

+++ src/ConsultaMoneda.java
@@ -7,7 +7,7 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 
 public class ConsultaMoneda {
-    public Moneda buscaMoneda (String tipoMoneda1, String tipoMoneda2, int cantidad) throws IOException, InterruptedException {
+    public Moneda buscaMoneda (String tipoMoneda1, String tipoMoneda2, double cantidad) throws IOException, InterruptedException {
         URI direccion = URI.create("https://v6.exchangerate-api.com/v6/032d311090a892e96db40c80/pair/"+tipoMoneda1+"/"+tipoMoneda2+"/"+cantidad);
 
         HttpClient client = HttpClient.newHttpClient();
